+++ src/app/styl/views/action-bar.styl
@@ -1,0 +1,102 @@
+#action-bar
+    width 100%
+    height: 60px
+    .actions-bar
+        width 100%
+        margin-top: 32px
+        height: 60px
+        .go-back
+            margin: 20px 40px
+            color: #FFF;
+            display: inline-block
+            cursor: pointer
+            float: left;
+            position: relative
+            i
+                vertical-align: bottom
+                font-size: 32px
+            span
+                font-family: $MainFontBold
+                margin-left 30px
+                text-transform uppercase
+                font-size: 32px
+
+        .toolbar
+            float: right
+            margin: 20px 40px 0 0
+            position relative
+            li
+                display: inline-block
+                padding: 10px
+                .material-icons
+                    color: #FFF
+                    font-size 1.3em
+                .health-icon
+                    &.Bad
+                        color: #d53f3f
+                    &.Medium
+                        color: #ece523
+                    &.Good
+                        color: #a3e147
+                    &.Excellent
+                        color: #2ad942
+            .toggles-container
+                  display: flex
+                  flex: 2
+
+                .favourites-toggle
+                    &:hover
+                        &:before
+                            opacity: 0
+                        &:after
+                            opacity: 1
+
+
+                    &.selected
+                        height: 18px
+                        position: relative
+
+                        &:before
+                            content: "\f004"
+                            font-family: FontAwesome
+                            position: absolute
+                            font-size: 18px
+                            color: $FavoriteColor
+                            left: 0
+                            opacity: 1
+                            transition: opacity 0.5s
+
+                        &:after
+                            content: "\f004"
+                            font-family: FontAwesome
+                            position: absolute
+                            font-size: 18px
+                            color: #FFF
+                            left: 0
+                            opacity: 0
+                            transition: opacity 0.5s
+
+                        &:hover
+                            &:before
+                                opacity: 0
+                            &:after
+                                opacity: 1
+
+                .watched-toggle
+
+                    &:hover
+                        &:before
+                            opacity: 0
+                        &:after
+                            opacity: 1
+
+
+                    &.selected
+                        height: 18px
+                        position: relative
+
+                        &:hover
+                            &:before
+                                opacity: 0
+                            &:after
+                                opacity: 1
