+++ src/app/styl/views/app.styl
@@ -30,6 +30,7 @@ $MovieListPadding = calc(100% - 67px)
 
 @import 'browser/list'
 @import 'browser/item'
+@import 'generic_detail'
 @import 'movie_detail'
 @import 'movie_error'
 @import 'show_detail'
