+++ src/app/styl/views/filter_bar.styl
@@ -1,4 +1,4 @@
-.filter-bar
+.filter-bar, .show-detail-header
     z-index: 10
     width: 100%
     opacity: 1
