+++ src/app/styl/views/movie_detail.styl
@@ -3,135 +3,9 @@
     flex-direction: row
     justify-content: flex-start
     flex-flow: row wrap
-    height: 100vh
-    width: 100vw
     fixed: bottom left
     user-select: none
-    backface-visibility: hidden
-    background-color: $BgColor1
-    position: fixed
-    z-index 10
-    .actions-bar
-        width 100%
-        margin-top: 32px
-        height: 60px
-        .go-back
-            margin: 20px 40px
-            color: #FFF;
-            display: inline-block
-            cursor: pointer
-            float: left;
-            position: relative
-            i
-                vertical-align: bottom
-                font-size: 32px
-            span
-                font-family: $MainFontBold
-                margin-left 30px
-                text-transform uppercase
-                font-size: 32px
-
-        .toolbar-movies
-            float: right
-            margin: 20px 40px 0 0
-            position relative
-            li
-                display: inline-block
-                padding: 10px
-                .material-icons
-                    color: #FFF
-                    font-size 1.3em
-                .health-icon
-                    &.Bad
-                        color: #d53f3f
-                    &.Medium
-                        color: #ece523
-                    &.Good
-                        color: #a3e147
-                    &.Excellent
-                        color: #2ad942
-            .toggles-container
-                  display: flex
-                  flex: 2
-
-                .favourites-toggle
-                    &:hover
-                        &:before
-                            opacity: 0
-                        &:after
-                            opacity: 1
-
-
-                    &.selected
-                        height: 18px
-                        position: relative
-
-                        &:before
-                            content: "\f004"
-                            font-family: FontAwesome
-                            position: absolute
-                            font-size: 18px
-                            color: $FavoriteColor
-                            left: 0
-                            opacity: 1
-                            transition: opacity 0.5s
-
-                        &:after
-                            content: "\f004"
-                            font-family: FontAwesome
-                            position: absolute
-                            font-size: 18px
-                            color: #FFF
-                            left: 0
-                            opacity: 0
-                            transition: opacity 0.5s
-
-                        &:hover
-                            &:before
-                                opacity: 0
-                            &:after
-                                opacity: 1
-
-                .watched-toggle
-
-                    &:hover
-                        &:before
-                            opacity: 0
-                        &:after
-                            opacity: 1
-
-
-                    &.selected
-                        height: 18px
-                        position: relative
-
-                        &:hover
-                            &:before
-                                opacity: 0
-                            &:after
-                                opacity: 1
-    .backdrop-overlay
-        width: 100vw
-        height: 100vh
-        position: absolute
-        opacity: 0.85
-        background-color: $BackColor_dark
-
-    .backdrop
-        width: 100vw
-        height: 100vh
-        position: absolute
-        opacity: 0
-        background-position-x 50%
-        background-position-y -33px
-        -webkit-filter: blur(15px) brightness(0.7)
-        background-repeat: no-repeat
-        background-size: cover
-
-        &.fadein
-            transition: opacity .3s ease-in;
-            opacity: 1
-
+    width: 100vw
     .poster-box
         height: 75vh
         width: 35vw
