+++ src/app/templates/generic-detail.tpl
@@ -1,0 +1,6 @@
+<div class="backdrop"></div>
+<div class="backdrop-overlay"></div>
+
+<div id="action-bar"></div>
+<div id="show-header"></div>
+<div id="detail-meat"></div>
