+++ src/app/templates/main-window.tpl
@@ -10,7 +10,7 @@
 <div id="keyboard-container"></div>
 <div id="content"></div>
 <div id="settings-container"></div>
-<div id="movie-detail"></div>
+<div id="details-container"></div>
 <div id="file-selector-container"></div>
 <div id="torrent-collection-container"></div>
 <div id="drop-mask"></div>
